+++ src/app/home/page.tsx
@@ -1,8 +1,8 @@
-import MapCard from "@component/map/MapCard";
-import Link from "next/link";
+import MapCard from "@component/map/cards/MapCard";
 import { Card, CardHeader, CardBody, CardFooter } from "@nextui-org/card";
 import { headers, cookies } from 'next/headers'
 import LogUtility from "@utils/LoggingUtility";
+import ActivityNearYouCard from "@component/map/cards/ActivityNearYouCard";
 
 export const runtime = 'edge';
 
@@ -31,18 +31,9 @@ export default async function Home() {
           }
         }} token={cookieStore.get('token')?.value || ''} />
       </div>
-      <div className="w-1/3 flex flex-col">
+      <div className="w-1/3 flex flex-col gap-3">
         <div>
-          <Card className="py-4 w-full">
-            <CardHeader className="pb-0 pt-2 px-4 flex-col items-start">
-              <p className="text-tiny uppercase font-bold">Daily Mix</p>
-              <small className="text-default-500">12 Tracks</small>
-              <h4 className="font-bold text-large">Frontend Radio</h4>
-            </CardHeader>
-            <CardBody className="overflow-visible py-2">
-              open requests near you
-            </CardBody>
-          </Card>
+          <ActivityNearYouCard />
         </div>
         <div>
           <Card className="py-4 w-full">
