+++ src/components/map/widgets/CompassWidget.tsx
